+++ inst/doc/miceFast-intro.R
@@ -1,8 +1,8 @@
 ## ----setup, include=FALSE------------------------------------------------
 knitr::opts_chunk$set(echo = FALSE)
 
-## ----echo=TRUE-----------------------------------------------------------
-library(miceFast)
+## ----echo=TRUE,message=FALSE,warning=FALSE-------------------------------
+library(miceFast);library(data.table);library(magrittr)
 set.seed(123456)
 
 ## ----eval=FALSE,echo=TRUE------------------------------------------------
@@ -24,9 +24,8 @@ model$update_var(2,model$impute("lm_pred",2,5)$imputations)
 
 model$update_var(3,model$impute("lda",3,c(1,2))$imputations) 
 
-#Old slow syntax
-#model$update_var(4,rowMeans(sapply(1:10,function(x) 
-#  model$impute("lm_bayes",4,c(1,2,3))$imputations)))
+#Old slow syntax model$update_var(4,rowMeans(sapply(1:10,function(x) model$impute("lm_bayes",4,c(1,2,3))$imputations)))
+#New syntax - impute_N
 model$update_var(4,model$impute_N("lm_bayes",4,c(1,2,3),10)$imputations)
 
 #When working with 'Big Data'
@@ -36,9 +35,9 @@ model$update_var(4,model$impute_N("lm_bayes",4,c(1,2,3),10)$imputations)
 # That is why `update_var` could be used only ones for a certain column
 # check which variables was updated - inside the object
 model$which_updated()
-head(model$get_data(),4)
-head(data,4)
-head(mice::nhanes,4)
+head(model$get_data(),3)
+head(data,3)
+head(mice::nhanes,3)
 rm(model)
 
 ## ----echo=TRUE-----------------------------------------------------------
@@ -60,11 +59,11 @@ model$update_var(2,model$impute_N("lm_bayes",2,c(1,3,4,5,6),10)$imputations)
 #Printing data and retrieving an old order
 head(cbind(model$get_data(),model$get_g(),model$get_w())[order(model$get_index()),],4)
 
-head(airquality,4)
+head(airquality,3)
 
-head(cbind(model$get_data(),model$get_g(),model$get_w()),4)
+head(cbind(model$get_data(),model$get_g(),model$get_w()),3)
 
-head(cbind(data,groups,weights),4)
+head(cbind(data,groups,weights),3)
 
 rm(model)
 
@@ -73,7 +72,7 @@ rm(model)
 data = cbind(as.matrix(airquality[,-5]),intercept = 1,index = 1:nrow(airquality))
 weights = rgamma(nrow(data),3,3) # a numeric vector - positive values
 #groups = as.numeric(airquality[,5]) # a numeric vector not integers - positive values
-groups = as.numeric(sample(1:3,nrow(data),replace=T)) # a numeric vector not integers - positive values
+groups = as.numeric(sample(1:8,nrow(data),replace=T)) # a numeric vector not integers - positive values
 
 model = new(miceFast)
 model$set_data(data) # providing by a reference
@@ -103,7 +102,7 @@ rm(model)
 #str(mtcars)
 mtcars$cyl= factor(mtcars$cyl)
 mtcars$gear= factor(mtcars$gear)
-mtcars_mat = model.matrix(~.,mtcars)
+mtcars_mat = model.matrix.lm(~.,mtcars,na.action="na.pass")
 #str(mtcars_mat)
 
 ## ----echo=TRUE-----------------------------------------------------------
@@ -115,6 +114,11 @@ airquality2$Temp2 = airquality2$Temp**2
 airquality2_mat = as.matrix(airquality2)
 model = new(miceFast)
 model$set_data(airquality2_mat)
-
 as.vector(model$vifs(1,c(2,3,4,5,6,7)))
+
+data_DT = data.table(airquality2)
+# VIF for variables at 1,3,4 positions - you include a y position to consider its NA values
+as.vector(data_DT[,.(vifs=VIF(x=as.matrix(.SD),
+                posit_y=1,
+                posit_x=c(2,3,4,5,6,7)))][['vifs']])