+++ src/miceFast_class.cpp
@@ -297,7 +297,9 @@ Rcpp::List miceFast::impute(std::string s, int posit_y,arma::uvec posit_x){
   posit_x =  posit_x - 1;
   posit_y = posit_y - 1;
 
-  arma::colvec pred =  option_impute_multiple(s,posit_y,posit_x,1);
+  const int times = 1;
+
+  arma::colvec pred =  option_impute_multiple(s,posit_y,posit_x,times);
 
   //index
 
@@ -353,21 +355,28 @@ arma::colvec miceFast::option_impute_multiple(std::string s,int posit_y,arma::uv
 
 arma::colvec miceFast::impute_raw(std::string s, int posit_y,arma::uvec posit_x,int times){
 
-  arma::mat X = x.cols(posit_x);
-  arma::colvec Y = x.col(posit_y);
+  arma::uvec posit_y_uvec(1);
+  posit_y_uvec(0) = posit_y;
 
   index_full = miceFast::get_index_full(posit_y, posit_x);
   index_NA = miceFast::get_index_NA(posit_y, posit_x);
 
-  if(!Y.has_nan()){Rcpp::stop("There is no NA values for the dependent variable");}
+  if(!x.col(posit_y).has_nan()){Rcpp::stop("There is no NA values for the dependent variable");}
 
-  arma::mat X_full = X.rows(index_full);
-  arma::mat X_NA = X.rows(index_NA);
-  arma::colvec Y_full = Y.rows(index_full);
+  arma::mat X_full = x(index_full,posit_x);
+  arma::mat X_NA = x(index_NA,posit_x);
+  arma::colvec Y_full = x(index_full,posit_y_uvec);
+
+  arma::colvec pred = x(index_NA,posit_y_uvec);
+
+  if(!(index_NA.n_elem==0) && ((index_full.n_elem>15 && s=="lda")|| (index_full.n_elem>posit_x.n_elem && s!="lda"))){
 
   pfunc f = funMap[s];
-  arma::colvec pred = (*f)(Y_full,X_full,X_NA,times);
+  pred = (*f)(Y_full,X_full,X_NA,times);
 
+  }
+
+  arma::colvec Y = x.col(posit_y);
   Y.rows(index_NA) = pred;
 
   return Y;
@@ -382,18 +391,20 @@ arma::colvec miceFast::imputeby(std::string s, int posit_y,arma::uvec posit_x, i
     sortData_byg();
   }
 
-  arma::mat X = x.cols(posit_x);
-  arma::colvec Y = x.col(posit_y);
+  arma::uvec posit_y_uvec(1);
+  posit_y_uvec(0) = posit_y;
 
   index_full = miceFast::get_index_full(posit_y, posit_x);
   index_NA = miceFast::get_index_NA(posit_y, posit_x);
 
   //if(!Y.has_nan()){Rcpp::stop("There is no NA values for the dependent variable");}
 
-  if(!Y.has_nan()){Rcpp::stop("There is no NA values for the dependent variable");}
+  if(!x.col(posit_y).has_nan()){Rcpp::stop("There is no NA values for the dependent variable");}
   if(g.has_nan()){Rcpp::stop("There is NA values for the grouping variable");}
 
-  arma::uvec g_int = arma::conv_to<arma::uvec>::from(g);
+  arma::uvec g_int(N_rows);
+
+  g_int = arma::conv_to<arma::uvec>::from(g);
 
   //grouping variable
 
@@ -407,55 +418,57 @@ arma::colvec miceFast::imputeby(std::string s, int posit_y,arma::uvec posit_x, i
 
   //dividing data to NA and full
 
-  arma::mat X_full = X.rows(index_full);
-  arma::mat X_NA = X.rows(index_NA);
-  arma::colvec Y_full = Y.rows(index_full);
-
   arma::uvec g_full = g_int.elem(index_full);
   arma::uvec g_NA = g_int.elem(index_NA);
 
   //predictions container
 
-  arma::colvec pred_all(index_NA.n_elem);
+  arma::colvec pred_all =  x(index_NA,posit_y_uvec) ;
 
   //iter
 
   arma::uvec his_full = arma::hist(g_full,un);
-  arma::uvec ends_full = arma::cumsum(his_full)-1;
+  arma::uvec ends_full = arma::cumsum(his_full);
   arma::uvec starts_full = arma::shift(ends_full,1) + 1;
-  starts_full(0) = 0;
+  starts_full(0) = 1;
 
   arma::uvec his_NA = arma::hist(g_NA,un);
-  arma::uvec ends_NA = arma::cumsum(his_NA)-1;
+  arma::uvec ends_NA = arma::cumsum(his_NA);
   arma::uvec starts_NA = arma::shift(ends_NA,1) + 1;
-  starts_NA(0) = 0;
+  starts_NA(0) = 1;
 
-  unsigned int start = 0;
-  unsigned int end = 0;
+  //unsigned int start = 0;
+  //unsigned int end = 0;
 
   for(unsigned int  a=0;a<group;a++){
 
-    int ss_NA = starts_NA(a);
-    int ee_NA = ends_NA(a);
-    int ss_full = starts_full(a);
-    int ee_full = ends_full(a);
+    int ss_NA = (int)starts_NA(a) - 1L;
+    int ee_NA = (int)ends_NA(a) - 1L;
+    int ss_full = (int)starts_full(a) - 1L;
+    int ee_full = (int)ends_full(a) - 1L;
 
-    if((ss_NA < ee_NA) && (ss_full <= ee_full)){
+    if((ss_NA <= ee_NA) && (ss_full <= ee_full)){
 
-      arma::mat X_full_0 = X_full.rows(ss_full,ee_full) ;  // copying. It would be better to improve it to use reference
-      arma::mat X_NA_0 = X_NA.rows(ss_NA,ee_NA);  // copying. It would be better to improve it to use reference
-      arma::colvec Y_full_0 =  Y_full.rows(ss_full,ee_full);  // copying. It would be better to improve it to use reference
+      arma::mat X_full_0 = x(index_full.subvec(ss_full,ee_full),posit_x) ;
+      arma::mat X_NA_0 = x(index_NA.subvec(ss_NA,ee_NA),posit_x);
+      arma::colvec Y_full_0 =  x(index_full.subvec(ss_full,ee_full),posit_y_uvec) ;
+
+      unsigned int N_obs_chunk = X_full_0.n_rows;
+
+      if((N_obs_chunk<=15 && s=="lda")|| (N_obs_chunk<posit_x.n_elem && s!="lda")){ continue ;}
 
       arma::colvec pred =  (*fun)(Y_full_0,X_full_0,X_NA_0,times);
 
-      end = start + pred.n_elem - 1;
+      //end = start + pred.n_elem - 1;
 
-      pred_all.rows(start,end) = pred;
+      pred_all.rows(ss_NA,ee_NA) = pred;
 
-      start = end + 1;
+      //start = end + 1;
 
     }
   }
+
+  arma::colvec Y = x.col(posit_y);
 
   Y.rows(index_NA) = pred_all;
 
@@ -471,26 +484,33 @@ std::map<std::string, pfuncw> funMapw = {{"lm_pred",fastLm_weighted},
 
 arma::colvec miceFast::imputeW(std::string s,int posit_y,arma::uvec posit_x,int times){
 
-  arma::mat X = x.cols(posit_x);
-  arma::colvec Y = x.col(posit_y);
+  arma::uvec posit_y_uvec(1);
+  posit_y_uvec(0) = posit_y;
 
   index_full = miceFast::get_index_full(posit_y, posit_x);
   index_NA = miceFast::get_index_NA(posit_y, posit_x);
 
-  if(!Y.has_nan()){Rcpp::stop("There is no NA values for the dependent variable");}
+  if(!x.col(posit_y).has_nan()){Rcpp::stop("There is no NA values for the dependent variable");}
   if(w.has_nan()){Rcpp::stop("There is NA values for weights");}
   if(arma::any(w<0)){Rcpp::stop("There are ngative values for the weights variable");}
 
   //dividing data to NA and full
 
-  arma::mat X_full = X.rows(index_full);
-  arma::mat X_NA = X.rows(index_NA);
-  arma::colvec Y_full = Y.elem(index_full);
+  arma::mat X_full = x(index_full,posit_x);
+  arma::mat X_NA = x(index_NA,posit_x);
+  arma::colvec Y_full = x(index_full,posit_y_uvec);
   arma::colvec w_full = w.elem(index_full);
 
-  pfuncw f = funMapw[s];
-  arma::colvec pred = (*f)(Y_full,X_full,w_full,X_NA,times);
+  arma::colvec pred = x(index_NA,posit_y_uvec);
 
+  if(!(index_NA.n_elem==0) && ((X_full.n_rows>15 && s=="lda")|| (index_full.n_elem>posit_x.n_elem && s!="lda"))){
+
+  pfuncw f = funMapw[s];
+  pred = (*f)(Y_full,X_full,w_full,X_NA,times);
+
+  }
+
+  arma::colvec Y = x.col(posit_y);
   Y.rows(index_NA) = pred;
 
   return Y;
@@ -504,17 +524,19 @@ arma::colvec miceFast::imputebyW(std::string s,int posit_y,arma::uvec posit_x,in
     sortData_byg();
   }
 
-  arma::mat X = x.cols(posit_x);
-  arma::colvec Y = x.col(posit_y);
+  arma::uvec posit_y_uvec(1);
+  posit_y_uvec(0) = posit_y;
 
   index_full = miceFast::get_index_full(posit_y, posit_x);
   index_NA = miceFast::get_index_NA(posit_y, posit_x);
 
-  if(!Y.has_nan()){Rcpp::stop("There is no NA values for the dependent variable");}
+  if(!x.col(posit_y).has_nan()){Rcpp::stop("There is no NA values for the dependent variable");}
   if(w.has_nan()){Rcpp::stop("There is NA values for weights variable");}
   if(g.has_nan()){Rcpp::stop("There is NA values for the grouping variable");}
 
-  arma::uvec g_int = arma::conv_to<arma::uvec>::from(g);
+  arma::uvec g_int(N_rows);
+
+  g_int = arma::conv_to<arma::uvec>::from(g);
 
   //grouping variable
 
@@ -527,61 +549,62 @@ arma::colvec miceFast::imputebyW(std::string s,int posit_y,arma::uvec posit_x,in
 
   //dividing data to NA and full
 
-  arma::mat X_full = X.rows(index_full);
-  arma::mat X_NA = X.rows(index_NA);
-  arma::colvec Y_full = Y.rows(index_full);
-  arma::colvec w_full = w.rows(index_full);
-
   arma::uvec g_full = g_int.elem(index_full);
   arma::uvec g_NA = g_int.elem(index_NA);
 
   //predictions container
 
-  arma::colvec pred_all(index_NA.n_elem);
+  arma::colvec pred_all =  x(index_NA,posit_y_uvec) ;
 
   // start end
 
   //iter
 
   arma::uvec his_full = arma::hist(g_full,un);
-  arma::uvec ends_full = arma::cumsum(his_full)-1;
+  arma::uvec ends_full = arma::cumsum(his_full);
   arma::uvec starts_full = arma::shift(ends_full,1) + 1;
-  starts_full(0) = 0;
+  starts_full(0) = 1;
 
   arma::uvec his_NA = arma::hist(g_NA,un);
-  arma::uvec ends_NA = arma::cumsum(his_NA) - 1;
+  arma::uvec ends_NA = arma::cumsum(his_NA);
   arma::uvec starts_NA = arma::shift(ends_NA,1) + 1;
-  starts_NA(0) = 0;
+  starts_NA(0) = 1;
 
 
-  unsigned int start = 0;
-  unsigned int end = 0;
+  //unsigned int start = 0;
+  //unsigned int end = 0;
 
   for(unsigned int  a=0;a<group;a++){
 
-    int ss_NA = starts_NA(a);
-    int ee_NA = ends_NA(a);
-    int ss_full = starts_full(a);
-    int ee_full = ends_full(a);
+    int ss_NA = (int)starts_NA(a) - 1L;
+    int ee_NA = (int)ends_NA(a) - 1L;
+    int ss_full = (int)starts_full(a) - 1L;
+    int ee_full = (int)ends_full(a) - 1L;
 
-    if((ss_NA < ee_NA) && (ss_full <= ee_full)){
+    if((ss_NA <= ee_NA) && (ss_full <= ee_full)){
 
-    arma::mat X_full_0 = X_full.rows(ss_full,ee_full) ;  // copying. It would be better to improve it to use reference
-    arma::mat X_NA_0 = X_NA.rows(ss_NA,ee_NA);  // copying. It would be better to improve it to use reference
-    arma::colvec Y_full_0 =  Y_full.rows(ss_full,ee_full);  // copying. It would be better to improve it to use reference
-    arma::colvec w_full_0 =  w_full.rows(ss_full,ee_full); // copying. It would be better to improve it to use reference
+    arma::mat X_full_0 = x(index_full.subvec(ss_full,ee_full),posit_x) ;
+    arma::mat X_NA_0 = x(index_NA.subvec(ss_NA,ee_NA),posit_x);
+    arma::colvec Y_full_0 =  x(index_full.subvec(ss_full,ee_full),posit_y_uvec) ;
+    arma::colvec w_full_0 =  w(index_full.subvec(ss_full,ee_full));
+
+    unsigned int N_obs_chunk = X_full_0.n_rows;
+
+    if((N_obs_chunk<=15 && s=="lda")|| (N_obs_chunk<posit_x.n_elem && s!="lda")){ continue ;}
 
     arma::colvec pred =  (*fun)(Y_full_0,X_full_0,w_full_0,X_NA_0,times);
 
-    end = start + pred.n_elem - 1;
+    //end = start + pred.n_elem - 1;
 
-    pred_all.rows(start,end) = pred;
+    pred_all.rows(ss_NA,ee_NA) = pred;
 
-    start = end + 1;
+    //start = end + 1;
 
     }
 
   }
+
+  arma::colvec Y = x.col(posit_y);
 
   Y.rows(index_NA) = pred_all;
 
@@ -598,6 +621,7 @@ RCPP_MODULE(miceFast){
     //.field("w",&miceFast::w)
     //.field("updated",&miceFast::updated)
     //.field("sorted",&miceFast::sorted)
+
     .method("get_data", &miceFast::get_data)
     .method("get_w", &miceFast::get_w)
     .method("get_g", &miceFast::get_g)